requirements.txt
Django==3.2.18
mysqlclient==2.1.1
djangorestframework==3.14.0
